+++ modules/cmsadmin/src/apis/AdminController.php
@@ -17,6 +17,8 @@ class AdminController extends \admin\base\RestController
 {
     public function actionDataBlocks()
     {
+        $favs = Yii::$app->adminuser->identity->setting->get("blockfav", []);
+        
         $groups = [];
         foreach (BlockGroup::find()->asArray()->all() as $group) {
             $blocks = [];
@@ -29,6 +31,7 @@ class AdminController extends \admin\base\RestController
                     'id' => $block['id'],
                     'name' => $obj->name(),
                     'full_name' => $obj->getFullName(),
+                    'favorized' => array_key_exists($block['id'], $favs),
                 ];
             }
 
@@ -36,12 +39,33 @@ class AdminController extends \admin\base\RestController
                 continue;
             }
             
+            $group['is_fav'] = 0;
+            $group['toggle_open'] = (int) Yii::$app->adminuser->identity->setting->get("togglegroup.{$group['id']}", 1);
             $groups[] = [
                 'group' => $group,
                 'blocks' => $blocks,
             ];
         }
 
+        if (!empty($favs)) {
+
+            $favblocks = [];
+            foreach ($favs as $fav) {
+                $favblocks[] = $fav;
+            }
+            
+            array_unshift($groups, [
+                'group' => [
+                    'toggle_open' => (int) Yii::$app->adminuser->identity->setting->get("togglegroup.99999", 1),
+                    'id' => '99999',
+                    'is_fav' => 1,
+                    'name' => 'Favorites',
+                    'identifier' => 'favs',
+                ],
+                'blocks' => $favblocks,
+            ]);
+        }
+        
         return $groups;
     }